ActionMenu and ATBTT_PlayerController updated to our needs.
The ActionMenu has been hooked into the ATBTT_PlayerController so that
the buttons on the HUD actually carry out the actions specified. The
movement block in the ATBTT_PlayerController has been converted to
separate functions that are called in the ActionMenu on button clicks.

Attempted to set Tab as the keybind to switch between pawns but was
unsuccessful. Will look into this again later.